Added Invoke-remotescan and collect-logdata
The invoke-remotescan allows for faster scanning, by executing scan instances on the remote machine in parallel
The Collect-logdata aggregates the created logfiles after all scans are finished and places them in a central directory.

You can always append deletion of the files on the local machine. - or alternatively, change the remotescan logfile location to a network share. 
This hasn't been done in order to make the script usable on DMZ and standalone servers without a domain join.